Added new improved battery script
The script now checks if the low_battery script is running, the status of the battery and according to that it will restart the script. E.G.: once the 3 notification expire, unless the user will restart xorg the script won't work anymore.
With this new script isntead I check that battery is not charging and it's above the low battery level and the script for notification is not running already. If these 3 condition verify it will launch again the script and the user won't need to restart xorg